dogstatsd-metric-exporter/example/requirements.txt
flask<1.2
opentelemetry-sdk
opentelemetry-api
datadog
